requirements.txt
neural-renderer-pytorch==1.1.3
numpy==1.15.4
opencv-python==3.4.4.19
scikit-image==0.14.1
scipy==1.2.0
tqdm==4.19.9
